terraform/vault/requirements.txt
PyYAML==3.13
terraform-external-data==0.0.6
